refactoring generate_density_utils.py imports
